+++ geotrek/common/parsers.py
@@ -1167,6 +1167,8 @@ class GeotrekAggregatorParser:
         "TouristicEvent": ("geotrek.tourism.parsers", "GeotrekTouristicEventParser"),
         "Signage": ("geotrek.signage.parsers", "GeotrekSignageParser"),
         "Infrastructure": ("geotrek.infrastructure.parsers", "GeotrekInfrastructureParser"),
+        "Site": ("geotrek.outdoor.parsers", "GeotrekSiteParser"),
+        "Course": ("geotrek.outdoor.parsers", "GeotrekCourseParser"),
     }
 
     invalid_model_topology = ['Trek', 'POI', 'Service', 'Signage', 'Infrastructure']